requirements.txt
django~=4.2
django-redis
psycopg2
djangorestframework